refactor: Post Head 분리, Post 상세 페이지에서 재사용 가능
